server/clientpackets/C_ItemUSe.java 	((_random.nextInt(99) + 1) 改為 RandomArrayList.getArray100List() 差1%機率卻影響很大效能 	一方面也是搞不懂正常機率都採100% 為何麼這裡卻要搞個 (0~98)+1 => 1~99 疑似日版筆誤 	修正多處之前更改時所導致的筆誤(魚類相關) server/clientpackets/C_LeaveClan.java && C_War.java 	應該部會有人血盟成員超過 32767 吧 變數改為short類型 32為原以下CPU影響較巨 server/model/ElementalStoneGenerator.java 	正常來說沒有人會設超過32767顆元素石吧 變數改為short類型 32為原以下CPU影響較巨 	重生時間多+0 (非必要性修正 只是覺得官版的時間很長) 針對model&slientpackets跟目錄做詳細檢查 測底改為隨機矩陣的使用 && 修正部分早期較差的使用方法
git-svn-id: http://www.castiello-tsai.org:8523/svn/L1JTW@180 96aee97a-5c11-ed42-8518-d7f33e96f5f9

+++ src/l1j/server/server/clientpackets/C_BanClan.java
@@ -46,7 +46,7 @@ public class C_BanClan extends ClientBasePacket {
 		L1Clan clan = L1World.getInstance().getClan(pc.getClanname());
 		if (clan != null) {
 			String clanMemberName[] = clan.getAllMembers();
-			int i;
+			short i;
 			if (pc.isCrown() && pc.getId() == clan.getLeaderId()) { // 君主、かつ、血盟主
 				for (i = 0; i < clanMemberName.length; i++) {
 					if (pc.getName().toLowerCase().equals(s.toLowerCase())) { // 君主自身
